requirements.txt
fastapi==0.115.0
langchain_core==0.3.5
langchain_openai==0.2.0
Pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3